Invoke xsel with -i in case stdout is not a term
Previously, xsel was invoked with no flags, causing it to output the
current selection if stdout was not a terminal, rather than to read data
from stdin.

This could cause unexpected results when attempting to use imgur.sh from
within other scripts, as stdout could be polluted with unexpected
data.